move searchquery storage to settings
