+++ allure-generator/src/main/javascript/components/node-search/NodeSearchView.js
@@ -7,13 +7,21 @@ import {View} from 'backbone.marionette';
 class NodeSearchView extends View {
     template = template;
 
-    initialize({onSearch}) {
-        this.onSearch = onSearch;
+    initialize({settings}) {
+        this.settings = settings;
+    }
+
+    onRender() {
+        this.$('input').val(this.settings.getSearchQuery());
     }
 
     @on('input input')
     onChangeSorting(e) {
-        this.onSearch(e.target.value);
+        this.settings.setSearchQuery(e.target.value);
+    }
+
+    close() {
+        this.settings.setSearchQuery('');
     }
 
 }
